Tidy up a few minor things for compiler

+++ board3d/gtkcolour3d.c
@@ -172,9 +172,9 @@ static void UpdateColourPreview(void)
 	gtk_color_button_get_array(GTK_COLOR_BUTTON(pcpSpecular), specular);
 
 	tempTexture = col3d.textureInfo;	/* Remeber texture, as setupmat resets it */
-	SetupMat(&col3d, ambient[0], ambient[1], ambient[2],
-		diffuse[0], diffuse[1], diffuse[2],
-		specular[0], specular[1], specular[2],
+	SetupMat(&col3d, (float)ambient[0], (float)ambient[1], (float)ambient[2],
+		(float)diffuse[0], (float)diffuse[1], (float)diffuse[2],
+		(float)specular[0], (float)specular[1], (float)specular[2],
 		(int)padjShine->value, opacityValue);
 	col3d.textureInfo = tempTexture;
